+++ spec/GitHub/SearchSpec.hs
@@ -2,10 +2,10 @@
 {-# LANGUAGE TemplateHaskell   #-}
 module GitHub.SearchSpec where
 
-import Prelude        ()
+import Prelude ()
 import Prelude.Compat
 
-import Data.Aeson.Compat  (eitherDecodeStrict)
+import Data.Aeson         (eitherDecodeStrict)
 import Data.FileEmbed     (embedFile)
 import Data.Proxy         (Proxy (..))
 import Data.String        (fromString)
@@ -14,7 +14,8 @@ import Test.Hspec         (Spec, describe, it, pendingWith, shouldBe)
 
 import qualified Data.Vector as V
 
-import GitHub.Data             (Auth (..), Issue (..), mkId, IssueState (..))
+import GitHub.Data
+       (Auth (..), Issue (..), IssueNumber (..), IssueState (..), mkId)
 import GitHub.Endpoints.Search (SearchResult (..), searchIssues')
 
 fromRightS :: Show a => Either a b -> b
@@ -40,13 +41,13 @@ spec = do
 
       let issue1 = issues V.! 0
       issueId issue1 `shouldBe` mkId (Proxy :: Proxy Issue) 123898390
-      issueNumber issue1 `shouldBe` 130
+      issueNumber issue1 `shouldBe` IssueNumber 130
       issueTitle issue1 `shouldBe` "Make test runner more robust"
       issueState issue1 `shouldBe` StateClosed
 
       let issue2 = issues V.! 1
       issueId issue2 `shouldBe` mkId (Proxy :: Proxy Issue) 119694665
-      issueNumber issue2 `shouldBe` 127
+      issueNumber issue2 `shouldBe` IssueNumber 127
       issueTitle issue2 `shouldBe` "Decouple request creation from execution"
       issueState issue2 `shouldBe` StateOpen