+++ src/GitHub/Data/Milestone.hs
@@ -40,3 +40,49 @@ instance FromJSON Milestone where
         <*> o .: "url"
         <*> o .: "created_at"
         <*> o .: "state"
+
+data NewMilestone = NewMilestone
+    { newMilestoneTitle       :: !Text
+    , newMilestoneState       :: !Text
+    , newMilestoneDescription :: !(Maybe Text)
+    , newMilestoneDueOn       :: !(Maybe UTCTime)
+    }
+  deriving (Show, Data, Typeable, Eq, Ord, Generic)
+
+instance NFData NewMilestone where rnf = genericRnf
+instance Binary NewMilestone
+
+
+instance ToJSON NewMilestone where
+    toJSON (NewMilestone title state desc due) = object $ filter notNull
+        [ "title"       .= title
+        , "state"       .= state
+        , "description" .= desc
+        , "due_on"      .= due
+        ]
+      where
+        notNull (_, Null) = False
+        notNull (_, _)    = True
+
+data UpdateMilestone = UpdateMilestone
+  { updateMilestoneTitle       :: !(Maybe Text)
+  , updateMilestoneState       :: !(Maybe Text)
+  , updateMilestoneDescription :: !(Maybe Text)
+  , updateMilestoneDueOn       :: !(Maybe UTCTime)
+  }
+  deriving (Show, Data, Typeable, Eq, Ord, Generic)
+
+instance NFData UpdateMilestone where rnf = genericRnf
+instance Binary UpdateMilestone
+
+
+instance ToJSON UpdateMilestone where
+  toJSON (UpdateMilestone title state desc due) = object $ filter notNull
+      [ "title"       .= title
+      , "state"       .= state
+      , "description" .= desc
+      , "due_on"      .= due
+      ]
+    where
+      notNull (_, Null) = False
+      notNull (_, _)    = True